feat: Adding cross-compile option for target build
Updating Makefile to cross-compile the driver for a arm64 arch,
the driver will be running on an imx93-frdm board.
Also updating the platform driver code to support different
kernel versions (first tested on ubuntu kernel 6.14, the imx
board has a custom OS with 6.6.36 kernel), this was achieved
adding a compatible layer (compat.h)
Fixing a typo on the DT binding helper.